% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cleanup.R
\name{clean_up}
\alias{clean_up}
\title{Removes any NAs from a given dataset}
\usage{
clean_up(filepath)
}
\arguments{
\item{filepath}{\itemize{
\item ability to insert data
}}
}
\value{
A dataset with no lingering NA values present
}
\description{
Removes any NAs from a given dataset
}
