% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot.R
\name{jack_box}
\alias{jack_box}
\title{Generates a boxplot for a given dataset that contains a column for both weight and species_id}
\usage{
jack_box(filepath)
}
\arguments{
\item{filepath}{\itemize{
\item ability to insert data
}}
}
\value{
boxplot for weight among various species
}
\description{
Generates a boxplot for a given dataset that contains a column for both weight and species_id
}
