% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Histogram.R
\name{make_hist}
\alias{make_hist}
\title{Forms a histogram of a given dataset that contains a column for weight}
\usage{
make_hist(filepath)
}
\arguments{
\item{filepath}{\itemize{
\item ability to insert data
}}
}
\value{
histogram for weight
}
\description{
Forms a histogram of a given dataset that contains a column for weight
}
